requirements.txt
contractions==0.1.73
sumy
nltk
langdetect==1.0.9
faiss-cpu
sentence-transformers
deep_translator
accelerate
bitsandbytes
huggingface_hub
nltk>=3.9
pandas>=2.2.1
tqdm>=4.66.2
sumy>=0.9.0
numpy>=1.26.4
faiss-cpu>=1.7.4  
sentence-transformers>=2.5.1
PyYAML>=6.0.1
deep-translator>=1.11.9
fastapi>=0.110.0
pydantic>=2.7.1
langdetect>=1.0.9
uvicorn>=0.32.0  
python-multipart>=0.0.6  
fastapi
uvicorn
pydantic
pandas
numpy
sentence-transformers
faiss-cpu
transformers
python-dotenv
deep-translator
python-multipart
